최종

+++ src/main/java/com/sparta/miniproject/Post/PostService.java
@@ -6,10 +6,16 @@ import com.sparta.miniproject.Post.ResponseDto.PostGetResponse;
 import com.sparta.miniproject.Post.ResponseDto.PostGetResponseDto;
 import com.sparta.miniproject.Post.ResponseDto.PostPostResponse;
 import com.sparta.miniproject.Post.ResponseDto.PostPostResponseDto;
-import com.sparta.miniproject.Post.TestUser.TestUser;
-import com.sparta.miniproject.Post.TestUser.TestUserRepository;
+import com.sparta.miniproject.model.Comment;
+import com.sparta.miniproject.model.Likes;
 import com.sparta.miniproject.model.Response;
+import com.sparta.miniproject.model.User;
+import com.sparta.miniproject.repository.CommentRepository;
+import com.sparta.miniproject.repository.LikeRepository;
+import com.sparta.miniproject.repository.UserRepository;
+import com.sparta.miniproject.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -21,13 +27,16 @@ import java.util.List;
 public class PostService
 {
     private final PostRepository postRepository;
-    private final TestUserRepository userRepository;
+    private final UserRepository userRepository;
+    private final LikeRepository likeRepository;
+    private final CommentRepository commentRepository;
 
-    public Response createPost(PostPostRequestDto requestDto)
+    @Transactional
+    public Response createPost(PostPostRequestDto requestDto, @AuthenticationPrincipal UserDetailsImpl userDetails)
     {
-        TestUser user = userRepository.findTestUserByNickname(requestDto.getNickname());
+        User user = userRepository.findByNickname(userDetails.getUser().getNickname()).orElse(null);
 
-        Post post = new Post(requestDto,user);
+        Post post = new Post(requestDto, user);
         postRepository.save(post);
 
         Response response = new Response();
@@ -46,11 +55,11 @@ public class PostService
 
         if ( postType.equals("time"))
         {
-            DBresponse = postRepository.findAllByOrderByCreatedAtDesc();
+            DBresponse = postRepository.findAllByOrderByCreatedAtAsc();
         }
         else if ( postType.equals("like"))
         {
-            DBresponse = postRepository.findAllByOrderByLikeCnt();
+            DBresponse = postRepository.findAllByOrderByLikeCntDesc();
         }
         else
         {
@@ -59,7 +68,10 @@ public class PostService
 
         for ( Post index : DBresponse)
         {
-            PostPostResponseDto postResponseDto = new PostPostResponseDto(index);
+            List<Comment> commentList = commentRepository.findAllByPost(index);
+            List<Likes> likesList = likeRepository.findAllByPost(index);
+
+            PostPostResponseDto postResponseDto = new PostPostResponseDto(index, likesList.size(), commentList.size());
             Response.add(postResponseDto);
 //            System.out.println(postResponseDto.getPostId());
 //            System.out.println(postResponseDto.getTitle());
@@ -81,10 +93,14 @@ public class PostService
     }
 
     @Transactional // SQL 쿼리가 일어나야 함을 스프링에게 알려줌
-    public Response updatePost(PostPutRequestDto requestDto, Long postId)
+    public Response updatePost(PostPutRequestDto requestDto, Long postId, @AuthenticationPrincipal UserDetailsImpl userDetails)
     {
         Post getPost= postRepository.findById(postId).orElseThrow(
                 ()-> new IllegalArgumentException("해당 포스트 없음") );
+
+        if (!getPost.getUser().getId().equals(userDetails.getUser().getId())) {
+            throw new IllegalArgumentException("작성자만 수정할수 있습니다.");
+        }
 
         getPost.update(requestDto);
 
@@ -95,8 +111,16 @@ public class PostService
 
     }
 
-    public Response deletePost(Long postId)
+    public Response deletePost(Long postId, @AuthenticationPrincipal UserDetailsImpl userDetails)
     {
+        Post post = postRepository.findById(postId).orElseThrow(
+                () -> new IllegalArgumentException("글이 존재하지 않습니다.")
+        );
+
+        if (!post.getUser().getId().equals(userDetails.getUser().getId())) {
+            throw new IllegalArgumentException("작성자만 삭제할수 있습니다.");
+        }
+
         postRepository.deleteById(postId);
 
         Response response = new Response();
@@ -105,12 +129,20 @@ public class PostService
         return response;
     }
 
-    public PostGetResponse getPost(Long postId)
+    public PostGetResponse getPost(Long postId, UserDetailsImpl userDetails)
     {
         Post getPost =  postRepository.findById(postId).orElseThrow(
                 ()->new IllegalArgumentException("해당 글이 존재하지 않습니다."));
 
-        PostGetResponseDto postGetResponseDto = new PostGetResponseDto(getPost);
+        Likes likes = likeRepository.findByUserAndPost(userDetails.getUser(), getPost).orElse(null);
+
+        List<Comment> commentList = commentRepository.findAllByPost(getPost);
+
+        List<Likes> likesList = getPost.getLikes();
+        boolean islike = false;
+        islike = likes != null;
+
+        PostGetResponseDto postGetResponseDto = new PostGetResponseDto(getPost, likesList.size(), islike, commentList.size());
 
         PostGetResponse postGetResponse = new PostGetResponse();
